src/migrations/requirements.txt
PyMySQL==1.1.1
yoyo-migrations==9.0.0
